# openapi.yaml
openapi: 3.0.3
info:
  title: Project webshop REST API
  description: A  RESTful API for managing products, categories, customers, addresses, orders, and suppliers.
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.webshop.com/v1
    description: Production server
  - url: https://sandbox.api.webshop.com/v1
    description: Sandbox/Testing server
security:
  - BearerAuth: []
paths:
  /products:
    get:
      operationId: listProducts
      security: []
      tags:
        - Products
      summary: List products
      description: Retrieve a list of products with optional filtering by category, supplier, or search term.
      parameters:
        - name: category_id
          in: query
          required: false
          description: Filter products by category ID
          schema:
            type: integer
        - name: supplier_id
          in: query
          required: false
          description: Filter products by supplier ID
          schema:
            type: integer
        - name: search
          in: query
          required: false
          description: Search text matching product name or description
          schema:
            type: string
      responses:
        '200':
          description: List of products retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '400':
          description: Error Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createProduct
      tags:
        - Products
      summary: Create a new product
      description: Add a new product to the catalog.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Error Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Error Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /products/{id}:
    get:
      operationId: getProductById
      security: []
      tags:
        - Products
      summary: Retrieve detailed product info
      description: Get complete information about a specific product by its ID.
      parameters:
        - name: id
          in: path
          required: true
          description: The product ID
          schema:
            type: integer
      responses:
        '200':
          description: Product retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Error Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: updateProduct
      tags:
        - Products
      summary: Update product details or stock quantity
      description: Modify details of an existing product or update its available stock count.
      parameters:
        - name: id
          in: path
          required: true
          description: The product ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Error Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Error Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Error Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteProduct
      tags:
        - Products
      summary: Delete a product
      description: Remove a product from the catalog by its ID.
      parameters:
        - name: id
          in: path
          required: true
          description: The product ID
          schema:
            type: integer
      responses:
        '204':
          description: Product deleted successfully
        '401':
          description: Error Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Error Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Provide your JWT bearer token in the format `Bearer <token>` to access protected endpoints.
  schemas:
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: INVALID_PARAMETER
        message:
          type: string
          example: The category_id field is required.
...
    Product:
      type: object
      required:
        - id
        - name
        - price
        - stock_quantity
        - category_id
        - supplier_id
      properties:
        id:
          type: integer
          example: 101
        name:
          type: string
          example: Wireless Noise-Canceling Headphones
        description:
          type: string
          nullable: true
          example: High-fidelity audio with active noise-canceling technology.
        price:
          type: number
          minimum: 0.01
          example: 199.99
        stock_quantity:
          type: integer
          minimum: 0
          example: 45
        category_id:
          type: integer
          example: 10
        supplier_id:
          type: integer
          example: 501
        created_at:
          type: string
          format: date-time
          example: '2026-09-02T01:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2026-09-02T01:00:00Z'
    ProductInput:
      type: object
      required:
        - name
        - price
        - stock_quantity
        - category_id
        - supplier_id
      properties:
        name:
          type: string
          example: Wireless Noise-Canceling Headphones
        description:
          type: string
          nullable: true
          example: High-fidelity audio with active noise-canceling technology.
        price:
          type: number
          minimum: 0.01
          example: 199.99
        stock_quantity:
          type: integer
          minimum: 0
          example: 45
        category_id:
          type: integer
          example: 10
        supplier_id:
          type: integer
          example: 501
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          example: Wireless Noise-Canceling Headphones
        description:
          type: string
          nullable: true
          example: Updated high-fidelity audio headphones description.
        price:
          type: number
          minimum: 0.01
          example: 189.99
        stock_quantity:
          type: integer
          minimum: 0
          example: 50
        category_id:
          type: integer
          example: 10
        supplier_id:
          type: integer
          example: 501
...